src/main/java/io/github/guimatech/hexagonal/application/usecases/GetCustomerByIdUseCase.java
package io.github.guimatech.hexagonal.application.usecases;

import io.github.guimatech.hexagonal.application.UseCase;
import io.github.guimatech.hexagonal.application.entities.CustomerId;
import io.github.guimatech.hexagonal.application.repositories.CustomerRepository;

import java.util.Objects;
import java.util.Optional;

public class GetCustomerByIdUseCase
        extends UseCase<GetCustomerByIdUseCase.Input, Optional<GetCustomerByIdUseCase.Output>> {

    private final CustomerRepository customerRepository;

    public GetCustomerByIdUseCase(final CustomerRepository customerRepository) {
        this.customerRepository = Objects.requireNonNull(customerRepository);
    }

    @Override
    public Optional<Output> execute(final Input input) {
        return customerRepository.customerOfId(CustomerId.with(input.id))
                .map(customer -> new Output(
                        customer.customerId().value().toString(),
                        customer.cpf().value(),
                        customer.email().value(),
                        customer.name().value()));
    }

    public record Input(String id) {}

    public record Output(String id, String cpf, String email, String name) {}
}
